Page loading fixes
Separated condition checks from a new `setReadyLoadGallery` flag to fix reload loop issue.
Gallery params (from either source) will now be stored in state variables, and referenced when the Gallery is triggered to load.
Changed settings in S3 (static website) and CloudFront (error pages) to bypass 404 errors on paths that do not have static pages on the server, and allow Router to handle the requests.
Also added a redirect back to the index if a bad URL path is requested.